iniciadas correções básicas
