+++ semana19/projeto/.gitignore
@@ -2,3 +2,4 @@ node_modules
 package-lock.json
 build
 .env
+.vscode
