+++ semana19/projeto/package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "ts-node-dev": "^1.1.8",
@@ -19,9 +20,14 @@
   },
   "dependencies": {
     "@types/knex": "^0.16.1",
+    "@types/node": "^16.11.4",
+    "@types/uuid": "^8.3.1",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
+    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "knex": "^0.95.11",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "uuid": "^8.3.2"
   }
 }
