+++ semana19/projeto/src/index.ts
@@ -1,29 +1,8 @@
-import express, {Express} from "express";
-import cors from "cors";
-import knex from "knex";
-import { AddressInfo } from "net";
+import app from './app'
+import generateId from './services/idGenerator'
 
-const app: Express = express();
+app.get('/')
 
-app.use(express.json());
-app.use(cors());
-
-export const connection = knex({
-	client: "mysql",
-	connection: {
-    host: process.env.DB_HOST,
-    port: 3306,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASS,
-    database: process.env.DB_NAME
-  }
-});
-
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in http://localhost: ${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
-});
+console.log(
+    generateId()
+)
