+++ semana19/projeto/src/services/hashManager.ts
@@ -1,0 +1,12 @@
+import { hashSync, compareSync, genSaltSync } from "bcryptjs";
+import { config } from "dotenv";
+
+config()
+
+export const generateHash = (
+    plainText: string
+): string => {
+    const rounds: number = Number (process.env.BCRYPT_COST)
+    const salt: string = genSaltSync(rounds)
+    return hashSync(plainText, salt)
+}
